+++ src/main/java/io/github/astrapi69/resourcebundle/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides a class for resolve configuration properties
+ */
+package io.github.astrapi69.resourcebundle.config;