+++ src/main/java/io/github/astrapi69/resourcebundle/file/namefilter/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides specific classes for properties an xml file names that implements the
+ * inteface java.io.FilenameFilter
+ */
+package io.github.astrapi69.resourcebundle.file.namefilter;