+++ src/main/java/io/github/astrapi69/resourcebundle/inspector/io/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides a class for reading the lines of a properties file
+ */
+package io.github.astrapi69.resourcebundle.inspector.io;