+++ src/main/java/io/github/astrapi69/resourcebundle/locale/ResourceBundleExtensions.java
@@ -29,9 +29,9 @@ import java.util.Locale;
 import java.util.MissingResourceException;
 import java.util.ResourceBundle;
 
+import io.github.astrapi69.check.Check;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
-import io.github.astrapi69.check.Check;
 
 /**
  * The class {@link ResourceBundleExtensions} holds extension methods for the class