+++ src/main/java/io/github/astrapi69/resourcebundle/properties/PropertiesKeyExtensions.java
@@ -27,9 +27,9 @@ package io.github.astrapi69.resourcebundle.properties;
 import java.util.Arrays;
 import java.util.List;
 
-import lombok.NonNull;
 import io.github.astrapi69.collection.list.ListExtensions;
 import io.github.astrapi69.collection.list.ListFactory;
+import lombok.NonNull;
 
 /**
  * The class {@link PropertiesKeyExtensions} provides methods for split properties keys to parts and