+++ src/main/java/io/github/astrapi69/resourcebundle/properties/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides a utility class for reading, loading and converting of properties files
+ */
+package io.github.astrapi69.resourcebundle.properties;