+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 ## Change log
 ----------------------
 
+Version 6.0
+-------------
+
+ADDED:
+
+- new module-info.java file for modularization
+- new libs.versions.toml file for new automatic catalog versions update
+- new class PropertiesExtensions that provides helper methods for Properties objects
+
+CHANGED:
+
+- update of JDK to version 17
+- update of gradle version to 8.10.2
+- update of gradle-plugin dependency com.github.ben-manes.versions.gradle.plugin to new version 0.51.0
+- update of gradle-plugin dependency com.diffplug.spotless:spotless-plugin-gradle to new version 7.0.0.BETA2
+- update of dependency lombok version to 1.18.34
+- update of dependency silly-collection to new minor version 28
+- update of dependency file-worker version to 17.4
+- update of test dependency test-object version to 9
+- update of test dependency testng version to 7.10.2
+- update of dependency silly-io version to 3
+- update of dependency json-extensions version to 3
+- remove of unused class NgxTranslateJsonFileToPropertiesFile
+- replaced obsolete package.html with package-info.java files
+
 Version 5
 -------------